Music search application based on a public Apple and wiki API.
1. Create a browser or native-app based application to server as a basic
music search app.
2. Search Screen
a. Allow customer to enter a term to search on.
b. Call the apple API and display all the returned track names, artist
name, album name, and image of album in a customer cell on a table,
c. Clicking a cell should navigate to the lyrics screen.
3. Lyrics screen
a. Display all the information from the previous cell plus the lyrics
returned from the wikia service.